Edited the Jupiter notebook
Displayed a head of n = 5 of the players.csv right above the mean values for age and played hours.

Made a new plot for a histogram of player ages.

Changed played hours vs age plot to be a scatterplot from a bar plot.

Added short descriptions to each of the two plots.